+++ R/plot_type_int.R
@@ -15,6 +15,7 @@ plot_type_int <- function(model,
                           geom.colors,
                           axis.title,
                           title,
+                          legend.title,
                           axis.lim,
                           case,
                           show.legend,
@@ -117,8 +118,21 @@ plot_type_int <- function(model,
     )
 
 
-    # select color palette
+    # evaluate dots-arguments
 
+    alpha <- .15
+    dodge <- .1
+    dot.alpha <- .5
+    log.y <- FALSE
+
+    add.args <- lapply(match.call(expand.dots = F)$`...`, function(x) x)
+    if ("alpha" %in% names(add.args)) alpha <- eval(add.args[["alpha"]])
+    if ("dodge" %in% names(add.args)) dodge <- eval(add.args[["dodge"]])
+    if ("dot.alpha" %in% names(add.args)) dot.alpha <- eval(add.args[["dot.alpha"]])
+    if ("log.y" %in% names(add.args)) log.y <- eval(add.args[["log.y"]])
+
+
+    # select color palette
     geom.colors <- col_check2(geom.colors, dplyr::n_distinct(dat$group))
 
 
@@ -133,7 +147,11 @@ plot_type_int <- function(model,
       jitter = jitter,
       use.theme = FALSE,
       case = case,
-      show.legend = show.legend
+      show.legend = show.legend,
+      dot.alpha = dot.alpha,
+      alpha = alpha,
+      dodge = dodge,
+      log.y = log.y
     )
 
     # set axis and plot titles
@@ -149,6 +167,10 @@ plot_type_int <- function(model,
     # set axis and plot titles
     if (!is.null(title))
       p <- p + ggplot::ggtitle(title)
+
+    # set axis and plot titles
+    if (!is.null(legend.title))
+      p <- p + labs(colour = legend.title)
 
     # set axis limits
     if (!is.null(axis.lim)) {